* This is a major update, affecting all existing projects. New projects are added. * The next update will set the stage for the beta release.
NOTE: A change log is now included for more complete update information.
            See ../Main/Source/Changelog.txt

In general:
1. Added Sandcastle.XsdDocs.dll, for XSD documentations.
	- Uses a customized version of the XML Schema Documenter by Immo Landwerth
	- Original source available at http://xsddoc.codeplex.com/ 
2. Added Sandcastle.Composition project: For plugin support.
	- Based on the current snapshot of the MEF Preview 6
	- Will be used for both .NET 3.5 and 4.0, renamed/signed to avoid confusion.
3. Added Sandcastle.Construction project: Manage proj without MSBuild dependency.
	A. Microsoft.Build.Construction implementations from the Mono Project
		- This is implemented by Leszek Ciesielski
	B. VC++ Projects equivalent of the Microsoft.Build.Constructions
		- Implemented to provide access to VC++ projects from VC++ 2003.
	C. Project Sections: Extracts relevant info from VS.NET projects and solutions
		- This presents a common interface to all projects.
		- The extracted information includes assembly/project references etc.
4. Completed work on most file formats and serialization of all the classes.
5. etc

+++ Main/Source/Components/ConceptualMathComponent.cs
@@ -206,7 +206,10 @@ namespace Sandcastle.Components
                         string mathFile  = _latexFormatter.ImageFile;
 
                         string outputFile = Path.Combine(_outputPath, mathFile);
-                        File.Move(_latexFormatter.ImagePath, outputFile);
+                        if (!File.Exists(outputFile))
+                        {
+                            File.Move(_latexFormatter.ImagePath, outputFile);
+                        }
 
                         //return String.Format("Equation|{0}|{1}:{2}{3}",
                         //    mathClass, equationNumber, _linkPath, mathFile);
@@ -225,7 +228,10 @@ namespace Sandcastle.Components
                         string mathFile  = _latexFormatter.ImageFile;
 
                         string outputFile = Path.Combine(_outputPath, mathFile);
-                        File.Move(_latexFormatter.ImagePath, outputFile);
+                        if (!File.Exists(outputFile))
+                        {
+                            File.Move(_latexFormatter.ImagePath, outputFile);
+                        }
 
                         //return String.Format("Equation|{0}|{1}:{2}{3}",
                         //    mathClass, equationNumber, _linkPath, mathFile);
@@ -246,7 +252,10 @@ namespace Sandcastle.Components
                         string mathFile  = _latexFormatter.ImageFile;
 
                         string outputFile = Path.Combine(_outputPath, mathFile);
-                        File.Move(_latexFormatter.ImagePath, outputFile);
+                        if (!File.Exists(outputFile))
+                        {
+                            File.Move(_latexFormatter.ImagePath, outputFile);
+                        }
 
                         _isNumbered = false;
                         //return String.Format("Equation|{0}|{1}:{2}{3}",
